fixed version in build script
