+++ package.json
@@ -13,7 +13,7 @@
     "analyze": "NODE_ENV=production webpack --config webpack/minified.config.js --json | webpack-bundle-size-analyzer",
     "build-all": "npm run clean && npm run build && npm run build-light && npm run build-minified && npm run build-minified-light",
     "precommit": "eslint src",
-    "predeploy": "npm run build-all",
+    "predeploy": "deploy=true npm run build-all",
     "deploy": "node webpack/deploy.js",
     "postdeploy": "git push origin master --follow-tags",
     "dry:deploy": "node webpack/deploy.js dry"
@@ -30,10 +30,10 @@
     "promise-polyfill": "^6.0.2",
     "video.js": "^6.9.0",
     "videojs-contextmenu": "^1.2.2",
+    "videojs-contrib-ads": "^6.2.0",
     "videojs-contrib-hls": "^5.14.1",
     "videojs-contrib-media-sources": "^4.4.2",
     "videojs-ima": "^1.5.1",
-    "videojs-contrib-ads": "^6.2.0",
     "videojs-per-source-behaviors": "^1.1.3",
     "webfontloader": "^1.6.28",
     "webpack-bundle-analyzer": "^2.9.1",
