+++ webpack/deploy.js
@@ -1,15 +1,7 @@
-const semver = require('semver');
+const ver = require('./versioning');
 const { execSync } = require('child_process');
-
-const CURRENT_VERSION = require('../package.json').version;
-const VALID_TAGS = ['edge', 'stable', 'dry'];
-
-const nextEdgeVersion = () => semver.inc(CURRENT_VERSION, 'prerelease', undefined, 'edge');
-const nextStableVersion = () => semver.inc(CURRENT_VERSION, 'patch');
-const nextVersion = (tag) => (tag === 'edge') ? nextEdgeVersion() : nextStableVersion();
-
-const versionCmd = (tag) => `npm version ${nextVersion(tag)}`;
-
+const tag = ver.extractTag();
+const versionCmd = () => `npm version ${VERSION}`;
 const publishCmd = (tag) => {
   let cmd = 'npm publish';
 
@@ -20,27 +12,12 @@ const publishCmd = (tag) => {
   return cmd;
 };
 
-const extractTag = () => {
-  let tag = process.argv[2];
-
-  if (!tag) {
-    return 'edge';
-  }
-
-  if (!VALID_TAGS.find((t) => t === tag)) {
-    throw new Error(`Invalid tag ${tag}. Valid tags are: ${VALID_TAGS.join(', ')}`);
-  }
-
-  return tag;
-};
-
-const tag = extractTag();
 
 if (tag === 'dry') {
-  console.log(`"edge" will deploy: "${nextEdgeVersion()}"`);
-  console.log(`"stable" will deploy: "${nextStableVersion()}"`);
+  console.log(`"edge" will deploy: "${ver.nextEdgeVersion()}"`);
+  console.log(`"stable" will deploy: "${ver.nextStableVersion()}"`);
 } else {
-  const cmd = `${versionCmd(tag)} && ${publishCmd(tag)}`;
+  const cmd = `${versionCmd()} && ${publishCmd(tag)}`;
   console.log(`Executing: "${cmd}" ...`);
   execSync(cmd);
 }
